isolating component scope using factory
