+++ templates/component-template/js/package.json
@@ -11,7 +11,7 @@
     "react": "16.10.2",
     "react-dom": "16.10.2",
     "react-scripts": "3.2.0",
-    "direflow-component": "1.0.23"
+    "direflow-component": "1.0.25"
   },
   "devDependencies": {
     "jest-environment-jsdom-fourteen": "0.1.0",
