+++ templates/component-template/js/src/index.js
@@ -2,6 +2,8 @@ import DireflowComponent from 'direflow-component';
 import { attributes, properties } from './properties';
 import App from './App';
 
-DireflowComponent.setAttributes(attributes);
-DireflowComponent.setProperties(properties);
-DireflowComponent.render(App, '%name-snake%');
+const direflowComponent = new DireflowComponent();
+
+direflowComponent.setAttributes(attributes);
+direflowComponent.setProperties(properties);
+direflowComponent.render(App, '%name-snake%');
