+++ templates/component-template/ts/package.json
@@ -11,7 +11,7 @@
     "@types/node": "12.7.8",
     "@types/react": "16.9.3",
     "@types/react-dom": "16.9.1",
-    "direflow-component": "1.0.23",
+    "direflow-component": "1.0.25",
     "react": "16.10.1",
     "react-dom": "16.10.1",
     "react-scripts": "3.1.2"
