+++ templates/project-template/ts/package.json
@@ -15,7 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "@types/jest": "24.0.15",
-    "direflow-project": "1.0.23",
+    "direflow-project": "1.0.25",
     "typescript": "^3.7.0",
     "tslint": "5.18.0",
     "tslint-config-airbnb": "5.11.1",
