+++ tslint.json
@@ -13,7 +13,8 @@
         "parameter"
       ],
       "severity": "warning"
-    }
+    },
+    "no-this-assignment": false,
   },
   "extends": ["tslint-config-airbnb", "tslint-config-silind"]
 }
